Release 2

+++ script.js
@@ -1,7 +1,7 @@
 
 // Get hard coded weather so no call need to be made. For testing only !!!!!
 function getWeatherDummy() {
-  console.log("In getWeatherDummy"); 
+  //console.log("In getWeatherDummy"); 
   
   
   $("#cityInfo").html("<h2>Detroit" + " (" + date + ") </h2>");
@@ -13,9 +13,43 @@ function getWeatherDummy() {
       
 }
 
+// Invalid input.... Assuming city name spelled wrong. Comes here for a 404 or 400 error.
+function invalidInput(cod,msg) {
+  //console.log("In invalidInput"); 
+    
+  $("#cityInfo").html("<h2>Sorry... App failed</h2>");
+  $("#cityInfo").append("<p>Looks like bad input</p>");  
+  $("#cityInfo").append("<p>Please check the spelling and Run again</p>");  
+  //$("#cityInfo").append("Error code is " + cod + " Error message is " + msg);
+
+  $("#flex1").empty();
+  $("#flex2").empty(); 
+  $("#flex3").empty();
+  $("#flex4").empty();
+  $("#flex5").empty();
+        
+}
+
+// Bad calls that do not include 404 or 400 error. 
+function systemError(cod,msg) {
+  console.log("In systemError"); 
+    
+  $("#cityInfo").html("<h2>SYSTEM ERROR</h2>");
+  $("#cityInfo").append("Error code is " + cod + " Error message is " + msg);
+  $("#cityInfo").append("<p>Please that losed Jim to look at it</p>");  
+
+  $("#flex1").empty();
+  $("#flex2").empty(); 
+  $("#flex3").empty();
+  $("#flex4").empty();
+  $("#flex5").empty();
+        
+}
+
+
 // Get todays weather for a city
 function getWeather() {
-  console.log("In getWeather");
+ //console.log("In getWeather");
   var WeatherUrl = "https://api.openweathermap.org/data/2.5/weather?q="
   var queryURL = WeatherUrl + city + apiKey + units;
 
@@ -30,9 +64,11 @@ function getWeather() {
      method: "GET"
    })
      // After data comes back from the request
-     .then(function(response) {
- //      console.log(queryURL);
-       console.log(response);
+     .then(function(response) {     
+ //      console.log(response);
+ //      console.log("!!!!!!!!!!!!!");
+ //      console.log(response.responseJSON.cod); 
+ //      console.log("!!!!!!!!!!!!!");
  //      console.log(response.name);
  //      console.log(response.main.temp);
  //      console.log(response.main.humidity);
@@ -45,20 +81,19 @@ function getWeather() {
  //      console.log("lat & lon");
  //      console.log("lon = " + lon);
  //      console.log("lat = " + lat);
-      console.log("icon num = " + iconNum);
-       //console.log(queryURL2);
+ //      console.log("icon num = " + iconNum);
+ //      console.log(queryURL2);
        
 
     //
     //   console.log(response.value);
-    //$("#cityInfo").html("<h1>" + response.name + " Weather Details</h1>");
     $("#cityInfo").html("<h2>" + response.name + " (" + date + ") " + "</h2>");
     
     var weatherIconUrl = weatherIconUrl1 + iconNum + weatherIconUrl3
-    console.log(weatherIconUrl)
+    //console.log(weatherIconUrl)
 
     var weatherIcon = $("<img>");
-    // Setting the catImage src attribute to imageUrl
+    // Setting the weatherIcon's src attribute to imageUrl
     weatherIcon.attr("src", weatherIconUrl);
     weatherIcon.attr("alt", "weather pic");
        
@@ -71,14 +106,26 @@ function getWeather() {
     getUVI();
     //getForecast();
  
-})        
- 
+}).catch(function(err){
+  console.log(err)
+  console.log("!!!!!!!!!!!!!");
+  console.log(err.responseJSON.cod);
+  console.log(err.responseJSON.message);
+  console.log("!!!!!!!!!!!!!");
+  if (err.responseJSON.cod == 404 || err.responseJSON.cod == 400)  {
+     invalidInput(err.responseJSON.cod, err.responseJSON.message);
+  } else {
+    systemError(err.responseJSON.cod, err.responseJSON.message);
+  } 
+
+})
+
       
 }
 
 // Get todays UVI for a city
 function getUVI() {
-  console.log("In getUVI");
+  //console.log("In getUVI");
   var uviUrl = "https://api.openweathermap.org/data/2.5/uvi?"
   var latStng = "lat=";
   var lonStng = "&lon=";
@@ -95,8 +142,7 @@ function getUVI() {
   })
     // After data comes back from the request
     .then(function(response) {
-      //console.log(queryURL2);
-        console.log(response);
+     // console.log(response);
      // console.log(response.value);
       $("#cityInfo").append("<p>UV Index: " + response.value + "</p>");
       getForecast();
@@ -110,10 +156,6 @@ function getForecast() {
   var ForcastUrl = "https://api.openweathermap.org/data/2.5/forecast?q="
   var forcastCall = ForcastUrl + city + apiKey + units; 
 
-  //var weatherIconUrl1 = "https://openweathermap.org/img/w/"
-  //var iconNum = "";
-  //var weatherIconUrl3 = ".png";
-
   $("#flex1").empty();
   $("#flex2").empty(); 
   $("#flex3").empty();
@@ -126,8 +168,8 @@ function getForecast() {
   })
     // After data comes back from the request
     .then(function(response) {
-  //    console.log(queryURL);
-      console.log(response);
+  //  console.log(queryURL);
+  //  console.log(response);
   
       var holdCurrent= "";
       var holdHighest = ""; 
@@ -162,7 +204,7 @@ function getForecast() {
             $("#flex1").append("<p>Humidity: " + response.list[i].main.humidity + " %" + "</p>");
             var dayOut = dayOut + 1;  
             holdHighest = holdCurrent;
-            console.log(holdHighest); 
+          //console.log(holdHighest); 
         } else if (holdCurrent != holdHighest && dayOut <5) { 
          //   console.log(response.list[i].main.temp)
          //   console.log(response.list[i].main.humidity)
@@ -178,37 +220,65 @@ function getForecast() {
         
       }
   
+      doButtonAndStorage();
    })  
-
+  
 }
-
-
 
 
 // Get incoming city & build a button for that city 
 function getInput() {
-  //console.log("In getInput");
+//  console.log("In getInput");
   //var inCity = document.getElementById(cityInput).value;
-  var incoming = document.forms["cityForm"]["cityName"].value
+  var incoming = document.forms["cityForm"]["cityName"].value;
+   
+
   //console.log(incoming);
   city = incoming;
+  //var incomingLen = city.length;
 
-  var a = $("<button>");
-  // Adding a class
-  a.addClass("City");
-  // Adding a data-attribute with a value of the movie at index i
-  a.attr("data-name", incoming);
-  // Providing the button's text with a value of the movie at index i
-  a.text(incoming);
-  // Adding the button to the HTML
-  $("#buttons-view").append(a)
+  //console.log("**** input  = " + incoming);
+  //console.log("**** lenght = " + incomingLen);
 
-  var x = citiesEntered.push(incoming);
-
-  //console.log(citiesEntered);
-  citiesEnteredJS = JSON.stringify(citiesEntered)
-  localStorage.setItem("Cites", citiesEnteredJS);
   getWeather() 
+}
+
+  function doButtonAndStorage() {
+
+    //console.log("In doButtonAndStorage");
+      var exists = false;
+
+      for (i=0; i <= (citiesEntered.length - 1); i++) {
+       // console.log(city.toUpperCase()); 
+       // console.log(citiesEntered[i].toUpperCase()); 
+        if (city.toUpperCase() === citiesEntered[i].toUpperCase()) {
+          var exists = true; 
+        }  
+      }  
+
+      //console.log(exists);
+
+      if (! exists) {
+      //  continue;
+      //} else {  
+          var a = $("<button>");
+          // Adding a class
+          a.addClass("City");
+          // Adding a data-attribute with a value of the movie at index i
+          a.attr("data-name", city);
+          // Providing the button's text with a value of the movie at index i
+          a.text(city);
+          // Adding the button to the HTML
+          $("#buttons-view").append(a)
+
+          var x = citiesEntered.push(city);
+
+          //console.log(citiesEntered);
+          citiesEnteredJS = JSON.stringify(citiesEntered)
+          localStorage.setItem("Cites", citiesEnteredJS);
+      }
+
+
 }
 
 //Goes to local storage & see if cities are out there. If there are build buttons.
@@ -246,6 +316,21 @@ function getLocalStorage() {
 
   } //End fuction
 
+  // Get hard coded weather so no call need to be made. For testing only !!!!!
+function clearList() {
+
+ //console.log("In clearList"); 
+
+  var value = localStorage.length; 
+  localStorage.clear("cities"); 
+
+  $("#buttons-view").empty(0)
+  citiesEntered.length = 0;
+
+//  console.log(citiesEntered);
+    
+}
+
 //PROGRAM STARTS HERE !!!!!!!!
 
 //Global Vars
@@ -275,7 +360,7 @@ var dateTime = mm + "/" + dd + "/" + yyyy + " " + hh + ":" + min;
 var date = mm + "/" + dd + "/" + yyyy;
 var todayDate = yyyy + "-" + mm + "-" + dd;
 
-console.log(todayDate);
+//console.log(todayDate);
 
 var apiKey = "&appid=073c4a93eb9c28d272af822a568d0d53";
 var units = "&units=imperial";
@@ -285,13 +370,17 @@ var long = "";
 
 getLocalStorage();
 
-console.log(citiesEntered.length);
+//console.log(citiesEntered.length);
 
 
 // Button listeners to start the process
 
 $("#btnCity").on("click", function () {
   getInput();
+});
+
+$("#btnClear").on("click", function () {
+  clearList();
 });
